feat(wayland): Update urgency handling
- Urgency is now in both xdg and xwayland structs
- Python handles the urgent status of windows as opposed to C
- Add missing urgency setter for type checking
- Share focus_on_window_activation logic between wayland and x11
- Wayland urgency is now handled based on focus_on_window_activation
- Create floates_states.py to prevent circular dependecy
