+++ test/test_hook.py
@@ -588,6 +588,9 @@ def test_client_urgent_hint_changed(manager_nospawn):
     manager_nospawn.start(ClientUrgentHintChangedConfig)
     manager_nospawn.test_window("Test Client", urgent_hint=True)
     assert_window(manager_nospawn, "Test Client")
+    # Get urgency of the window
+    _, urgent = manager_nospawn.c.eval("list(self.windows_map.values())[0].urgent")
+    assert urgent == "True"
 
 
 class CallLayoutGroup:
